# AoT optimization: trim out unused code